+++ src/main/java/uk/gov/laa/pfla/auth/service/services/UserService.java
@@ -2,6 +2,7 @@ package uk.gov.laa.pfla.auth.service.services;
 
 import com.microsoft.graph.core.ClientException;
 import com.microsoft.graph.models.User;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.oauth2.client.OAuth2AuthorizedClient;
 import org.springframework.stereotype.Component;
 import org.springframework.util.StringUtils;
@@ -9,52 +10,49 @@ import uk.gov.laa.pfla.auth.service.beans.UserDetails;
 import uk.gov.laa.pfla.auth.service.exceptions.AuthUserNotFoundException;
 import uk.gov.laa.pfla.auth.service.exceptions.UserServiceException;
 import uk.gov.laa.pfla.auth.service.graph.GraphClientHelper;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 /**
  * Component to encapsulate the retrieval of user information from an
  * external directory (Microsoft Graph)
  */
 @Component
+@Slf4j
 public class UserService {
 
-  public static final Logger log = LoggerFactory.getLogger(UserService.class);
+    private final GraphClientHelper graphClientHelper;
 
-  private final GraphClientHelper graphClientHelper;
-
-  public UserService(GraphClientHelper graphClientHelper) {
-    this.graphClientHelper = graphClientHelper;
-  }
-
-  /**
-   * Retrieve the details for the currently authenticated User, based on the
-   * supplied OAuth2AuthorizedClient.
-   *
-   * @param client a client which is authorized to query user details from the external directory
-   * @return a populated UserDetails
-   * @throws UserServiceException if an error, or no details, are returned from the external directory
-   */
-  public UserDetails getUserDetails(OAuth2AuthorizedClient client) throws UserServiceException {
-    User graphUser;
-    try {
-      graphUser = graphClientHelper.getGraphUserDetails(client);
-      if( graphUser == null || !StringUtils.hasText(graphUser.userPrincipalName) ) {
-        log.error("Null or insufficient User data returned from Graph: {}", graphUser);
-        throw new AuthUserNotFoundException("No user details returned from Graph");
-      }
-    } catch (ClientException e) {
-      throw new UserServiceException("Failed to retrieve Graph User", e);
+    public UserService(GraphClientHelper graphClientHelper) {
+        this.graphClientHelper = graphClientHelper;
     }
 
-    UserDetails userDetails = new UserDetails();
-    userDetails.setUserPrincipalName(graphUser.userPrincipalName);
-    userDetails.setGivenName(graphUser.givenName);
-    userDetails.setSurname(graphUser.surname);
-    userDetails.setPreferredName(graphUser.preferredName);
-    userDetails.setEmail(graphUser.mail);
+    /**
+     * Retrieve the details for the currently authenticated User, based on the
+     * supplied OAuth2AuthorizedClient.
+     *
+     * @param client a client which is authorized to query user details from the external directory
+     * @return a populated UserDetails
+     * @throws UserServiceException if an error, or no details, are returned from the external directory
+     */
+    public UserDetails getUserDetails(OAuth2AuthorizedClient client) throws UserServiceException {
+        User graphUser;
+        try {
+            graphUser = graphClientHelper.getGraphUserDetails(client);
+            if (graphUser == null || !StringUtils.hasText(graphUser.userPrincipalName)) {
+                log.error("Null or insufficient User data returned from Graph: {}", graphUser);
+                throw new AuthUserNotFoundException("No user details returned from Graph");
+            }
+        } catch (ClientException e) {
+            throw new UserServiceException("Failed to retrieve Graph User", e);
+        }
 
-    return userDetails;
-  }
+        UserDetails userDetails = new UserDetails();
+        userDetails.setUserPrincipalName(graphUser.userPrincipalName);
+        userDetails.setGivenName(graphUser.givenName);
+        userDetails.setSurname(graphUser.surname);
+        userDetails.setPreferredName(graphUser.preferredName);
+        userDetails.setEmail(graphUser.mail);
+
+        return userDetails;
+    }
 
 }
