+++ src/main/resources/application-dev.yml
@@ -14,20 +14,6 @@ gpfd:
       password: ${MOJFIN_DEV_WRITE_PASSWORD}
       driver-class-name: oracle.jdbc.OracleDriver
 
-logging:
-  level:
-    root: DEBUG
-    org:
-      springframework:
-        security: DEBUG
-        jdbc:
-          core:
-            JdbcTemplate: DEBUG
-            StatementCreatorUtils: DEBUG
-        transaction: DEBUG
-    #    org.springframework: INFO
-    #    org.springframework.security: INFO #TRACE
-    org.springframework.web.client.RestTemplate: DEBUG
 
 #SSL configuration
 server:
