+++ src/main/resources/application.yml
@@ -1,18 +1,12 @@
-# Azure AD and Graph config
-#spring:
-#  profile:
-#  cloud:
-#    azure:
-#      credential:
-#        client-secret: ${AZURE_CLIENT_SECRET}
-#        client-id: ${AZURE_CLIENT_ID}
-#      profile:
-#        tenant-id: ${AZURE_TENANT_ID}
-#
-#      active-directory:
-#        enabled: true
-#        authorization-clients:
-#          graph:
-#            scopes:
-#              - https://graph.microsoft.com/User.Read
-#        post-logout-redirect-uri: /
+logging:
+  level:
+    root: INFO
+    org:
+      springframework:
+        security: INFO
+        jdbc:
+          core:
+            JdbcTemplate: INFO
+            StatementCreatorUtils: INFO
+        transaction: INFO
+        web.client.RestTemplate: INFO
